requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
idna==3.6
numpy==1.26.2
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.4
sqlparse==0.4.4
urllib3==2.1.0
